requirements/requirements-3.12.txt
annotated-types==0.7.0
anyio==4.9.0
app-common-python==0.2.7
certifi==2025.4.26
click==8.2.1
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.17.0
greenlet==3.2.2
grpcio==1.73.0rc1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
inventory-grpc-clients-python-kessel-project==0.10.14
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pipenv==2024.4.1
platformdirs==4.3.6
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.0
protobuf==6.31.1
psycopg==3.2.9
psycopg-c==3.2.9
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.7
sentry-sdk==2.29.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
virtualenv==20.29.1
watchfiles==1.0.5
websockets==15.0.1
wheel==0.45.1
